# my-server-code-implementation/src/main/java/com/mycompany/api/MarketDataApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mycompany.api;

import com.mycompany.model.MarketDataCandlesGet200ResponseInner;
import com.mycompany.model.MarketDataOrderbookGet200Response;
import com.mycompany.model.MarketDataTradeHistoryGet200ResponseInner;
import com.mycompany.model.MarketDataV3OrderbookInstrumentFuturesDepthGet200Response;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-04T20:05:18.570459+05:30[Asia/Calcutta]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Public", description = "Public endpoints that don't require authentication")
public interface MarketDataApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_MARKET_DATA_CANDLES_GET = "/market_data/candles";
    /**
     * GET /market_data/candles : Get candlestick data
     * Returns candlestick bars for given pair
     *
     * @param pair Market pair from Market Details API (required)
     * @param interval Candlestick interval (required)
     * @param startTime Timestamp in milliseconds (optional)
     * @param endTime Timestamp in milliseconds (optional)
     * @param limit Number of candles to return (optional, default to 500)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "marketDataCandlesGet",
        summary = "Get candlestick data",
        description = "Returns candlestick bars for given pair",
        tags = { "Public" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = MarketDataCandlesGet200ResponseInner.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "SignatureAuth"),
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MarketDataApi.PATH_MARKET_DATA_CANDLES_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<MarketDataCandlesGet200ResponseInner>> marketDataCandlesGet(
        @NotNull @Parameter(name = "pair", description = "Market pair from Market Details API", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "pair", required = true) String pair,
        @NotNull @Parameter(name = "interval", description = "Candlestick interval", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "interval", required = true) String interval,
        @Parameter(name = "startTime", description = "Timestamp in milliseconds", in = ParameterIn.QUERY) @Valid @RequestParam(value = "startTime", required = false) @Nullable Long startTime,
        @Parameter(name = "endTime", description = "Timestamp in milliseconds", in = ParameterIn.QUERY) @Valid @RequestParam(value = "endTime", required = false) @Nullable Long endTime,
        @Max(value = 1000) @Parameter(name = "limit", description = "Number of candles to return", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "500") Integer limit
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"volume\" : 2.3021358869347655, \"high\" : 1.4658129805029452, \"low\" : 5.962133916683182, \"time\" : 0, \"close\" : 5.637376656633329, \"open\" : 6.027456183070403 }, { \"volume\" : 2.3021358869347655, \"high\" : 1.4658129805029452, \"low\" : 5.962133916683182, \"time\" : 0, \"close\" : 5.637376656633329, \"open\" : 6.027456183070403 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_MARKET_DATA_CANDLESTICKS_GET = "/market_data/candlesticks";
    /**
     * GET /market_data/candlesticks : Get Futures Candlesticks
     * Returns candlestick data for futures instruments
     *
     * @param pair Trading pair (required)
     * @param from Start timestamp (required)
     * @param to End timestamp (required)
     * @param resolution Candle resolution (required)
     * @param pcode Product code (f for futures, s for spot) (required)
     * @return Successful response with candlestick data (status code 200)
     */
    @Operation(
        operationId = "marketDataCandlesticksGet",
        summary = "Get Futures Candlesticks",
        description = "Returns candlestick data for futures instruments",
        tags = { "Public", "Futures" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response with candlestick data", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Object.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "SignatureAuth"),
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MarketDataApi.PATH_MARKET_DATA_CANDLESTICKS_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Object>> marketDataCandlesticksGet(
        @NotNull @Parameter(name = "pair", description = "Trading pair", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "pair", required = true) String pair,
        @NotNull @Parameter(name = "from", description = "Start timestamp", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "from", required = true) Long from,
        @NotNull @Parameter(name = "to", description = "End timestamp", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "to", required = true) Long to,
        @NotNull @Parameter(name = "resolution", description = "Candle resolution", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "resolution", required = true) String resolution,
        @NotNull @Parameter(name = "pcode", description = "Product code (f for futures, s for spot)", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "pcode", required = true) String pcode
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ \"{}\", \"{}\" ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_MARKET_DATA_ORDERBOOK_GET = "/market_data/orderbook";
    /**
     * GET /market_data/orderbook : Get order book
     * Returns a sorted list of bids and asks
     *
     * @param pair Market pair from Market Details API (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "marketDataOrderbookGet",
        summary = "Get order book",
        description = "Returns a sorted list of bids and asks",
        tags = { "Public" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MarketDataOrderbookGet200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "SignatureAuth"),
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MarketDataApi.PATH_MARKET_DATA_ORDERBOOK_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<MarketDataOrderbookGet200Response> marketDataOrderbookGet(
        @NotNull @Parameter(name = "pair", description = "Market pair from Market Details API", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "pair", required = true) String pair
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"asks\" : \"{}\", \"bids\" : \"{}\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_MARKET_DATA_TRADE_HISTORY_GET = "/market_data/trade_history";
    /**
     * GET /market_data/trade_history : Get trade history
     * Returns a sorted list of most recent 30 trades by default
     *
     * @param pair Market pair from Market Details API (required)
     * @param limit Number of trades to return (optional, default to 30)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "marketDataTradeHistoryGet",
        summary = "Get trade history",
        description = "Returns a sorted list of most recent 30 trades by default",
        tags = { "Public" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = MarketDataTradeHistoryGet200ResponseInner.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "SignatureAuth"),
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MarketDataApi.PATH_MARKET_DATA_TRADE_HISTORY_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<List<MarketDataTradeHistoryGet200ResponseInner>> marketDataTradeHistoryGet(
        @NotNull @Parameter(name = "pair", description = "Market pair from Market Details API", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "pair", required = true) String pair,
        @Max(value = 500) @Parameter(name = "limit", description = "Number of trades to return", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "30") Integer limit
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"p\" : 0.8008281904610115, \"q\" : 6.027456183070403, \"T\" : 1, \"m\" : true }, { \"p\" : 0.8008281904610115, \"q\" : 6.027456183070403, \"T\" : 1, \"m\" : true } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_MARKET_DATA_V3_ORDERBOOK_INSTRUMENT_FUTURES_DEPTH_GET = "/market_data/v3/orderbook/{instrument}-futures/{depth}";
    /**
     * GET /market_data/v3/orderbook/{instrument}-futures/{depth} : Get Futures Orderbook
     * Returns orderbook data for a futures instrument
     *
     * @param instrument Futures instrument name (required)
     * @param depth Orderbook depth (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "marketDataV3OrderbookInstrumentFuturesDepthGet",
        summary = "Get Futures Orderbook",
        description = "Returns orderbook data for a futures instrument",
        tags = { "Public", "Futures" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MarketDataV3OrderbookInstrumentFuturesDepthGet200Response.class))
            })
        },
        security = {
            @SecurityRequirement(name = "SignatureAuth"),
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = MarketDataApi.PATH_MARKET_DATA_V3_ORDERBOOK_INSTRUMENT_FUTURES_DEPTH_GET,
        produces = { "application/json" }
    )
    default ResponseEntity<MarketDataV3OrderbookInstrumentFuturesDepthGet200Response> marketDataV3OrderbookInstrumentFuturesDepthGet(
        @NotNull @Parameter(name = "instrument", description = "Futures instrument name", required = true, in = ParameterIn.PATH) @PathVariable("instrument") String instrument,
        @NotNull @Parameter(name = "depth", description = "Orderbook depth", required = true, in = ParameterIn.PATH) @PathVariable("depth") String depth
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"asks\" : [ [ \"50100.00\", \"0.3\" ], [ \"50100.00\", \"0.3\" ] ], \"bids\" : [ [ \"50000.00\", \"0.5\" ], [ \"50000.00\", \"0.5\" ] ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}