# spring-client/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: legal@coindcx.com
  description: "Welcome to the CoinDCX API!\n\nThe base URL for the API calls is `https://api.coindcx.com`,\
    \ base URL for some public endpoint is `https://public.coindcx.com`.\n\n## Important\
    \ Legal Notice\n\nThese API License Terms and Conditions govern the use of any\
    \ 'Market Data' and Application Programming Interface (API) of CoinDCX. \nBy accessing\
    \ any CoinDCX API, Market Data and/or any other information, service, or feature,\
    \ you shall be bound by these Terms.\n\n## Key Definitions\n\n- **Market Data**:\
    \ All data related to trading activity on CoinDCX platforms, including prices,\
    \ quantities of orders and transactions\n- **CoinDCX API**: Public API and/or\
    \ any data, information related to trading activity made available through https://docs.coindcx.com/\n\
    - **Algorithmic Trading**: Method using computer programs following defined instructions\
    \ or algorithms to place trades\n- **High-Frequency Trading**: Type of algorithmic\
    \ trading with high speeds, high turnover rates, and high order-to-trade ratios\n\
    \n## License Grant\n\nUsers are granted a **non-exclusive, non-transferable, non-assignable,\
    \ non-sublicensable, revocable, restricted license** for usage only.\n\nUsers\
    \ must NOT:\n- Alter, manipulate, or misrepresent any CoinDCX API or Market Data\n\
...
    \ a timestamp of when it was generated\n"
  license:
...
  title: CoinDCX API
...
servers:
...
security:
- ApiKeyAuth: []
  SignatureAuth: []
...
paths:
...
  /market_data/trade_history:
    get:
      description: Returns a sorted list of most recent 30 trades by default
      parameters:
      - description: Market pair from Market Details API
        example: B-BTC_USDT
        explode: true
        in: query
        name: pair
        required: true
        schema:
          type: string
        style: form
      - description: Number of trades to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 30
          maximum: 500
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/_market_data_trade_history_get_200_response_inner"
                type: array
          description: Successful response
      servers:
      - url: https://public.coindcx.com
      summary: Get trade history
      tags:
      - Public
      x-accepts:
      - application/json
  /market_data/orderbook:
    get:
      description: Returns a sorted list of bids and asks
      parameters:
      - description: Market pair from Market Details API
        example: B-BTC_USDT
        explode: true
        in: query
        name: pair
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_market_data_orderbook_get_200_response"
          description: Successful response
      servers:
      - url: https://public.coindcx.com
      summary: Get order book
      tags:
      - Public
      x-accepts:
      - application/json
  /market_data/candles:
    get:
      description: Returns candlestick bars for given pair
      parameters:
      - description: Market pair from Market Details API
        example: B-BTC_USDT
        explode: true
        in: query
        name: pair
        required: true
        schema:
          type: string
        style: form
      - description: Candlestick interval
        explode: true
        in: query
        name: interval
        required: true
        schema:
          enum:
          - 1m
          - 5m
          - 15m
          - 30m
          - 1h
          - 2h
          - 4h
          - 6h
          - 8h
          - 1d
          - 3d
          - 1w
          - 1M
          type: string
        style: form
      - description: Timestamp in milliseconds
        explode: true
        in: query
        name: startTime
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: Timestamp in milliseconds
        explode: true
        in: query
        name: endTime
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: Number of candles to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 500
          maximum: 1000
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/_market_data_candles_get_200_response_inner"
                type: array
          description: Successful response
      servers:
      - url: https://public.coindcx.com
      summary: Get candlestick data
      tags:
      - Public
      x-accepts:
      - application/json
  /market_data/v3/orderbook/{instrument}-futures/{depth}:
    get:
      description: Returns orderbook data for a futures instrument
      parameters:
      - description: Futures instrument name
        example: BTCUSDT
        explode: false
        in: path
        name: instrument
        required: true
        schema:
          type: string
        style: simple
      - description: Orderbook depth
        explode: false
        in: path
        name: depth
        required: true
        schema:
          enum:
          - "10"
          - "20"
          - "50"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_market_data_v3_orderbook__instrument__futures__depth__get_200_response"
          description: Successful response
      servers:
      - url: https://public.coindcx.com
      summary: Get Futures Orderbook
      tags:
      - Public
      - Futures
      x-accepts:
      - application/json
  /market_data/candlesticks:
    get:
      description: Returns candlestick data for futures instruments
      parameters:
      - description: Trading pair
        example: BTCUSDT
        explode: true
        in: query
        name: pair
        required: true
        schema:
          type: string
        style: form
      - description: Start timestamp
        explode: true
        in: query
        name: from
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - description: End timestamp
        explode: true
        in: query
        name: to
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - description: Candle resolution
        explode: true
        in: query
        name: resolution
        required: true
        schema:
          type: string
        style: form
      - description: "Product code (f for futures, s for spot)"
        explode: true
        in: query
        name: pcode
        required: true
        schema:
          enum:
          - f
          - s
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: object
                type: array
          description: Successful response with candlestick data
      servers:
      - url: https://public.coindcx.com
      summary: Get Futures Candlesticks
      tags:
      - Public
      - Futures
      x-accepts:
      - application/json
...
components:
  schemas:
...
    _market_data_trade_history_get_200_response_inner:
      example:
        p: 0.8008281904610115
        q: 6.027456183070403
        T: 1
        m: true
      properties:
        p:
          description: Trade price
          type: number
        q:
          description: Quantity
          type: number
        T:
          description: Timestamp
          type: integer
        m:
          description: Whether the buyer is market maker
          type: boolean
      type: object
    _market_data_orderbook_get_200_response:
      example:
        asks: "{}"
        bids: "{}"
      properties:
        bids:
          description: Buy orders sorted in descending order
          type: object
        asks:
          description: Sell orders sorted in ascending order
          type: object
      type: object
    _market_data_candles_get_200_response_inner:
      example:
        volume: 2.3021358869347655
        high: 1.4658129805029452
        low: 5.962133916683182
        time: 0
        close: 5.637376656633329
        open: 6.027456183070403
      properties:
        time:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      type: object
    _market_data_v3_orderbook__instrument__futures__depth__get_200_response:
      example:
        asks:
        - - "50100.00"
          - "0.3"
        - - "50100.00"
          - "0.3"
        bids:
        - - "50000.00"
          - "0.5"
        - - "50000.00"
          - "0.5"
      properties:
        bids:
          description: Buy orders
          items:
            description: "Price and quantity pair [price, quantity]"
            example:
            - "50000.00"
            - "0.5"
            items:
              type: string
            type: array
          type: array
        asks:
          description: Sell orders
          items:
            description: "Price and quantity pair [price, quantity]"
            example:
            - "50100.00"
            - "0.3"
            items:
              type: string
            type: array
          type: array
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: API key for authentication
      in: header
      name: X-AUTH-APIKEY
      type: apiKey
    SignatureAuth:
      description: HMAC-SHA256 signature of the request payload
      in: header
      name: X-AUTH-SIGNATURE
      type: apiKey